Unify shared-types files
