+++ src/components/PreviewModes/PreviewModes.tsx
@@ -3,7 +3,7 @@ import type { ValueBase, AriaLabelingProps } from "@react-types/shared";
 import { Radio, RadioGroup, Icon, ComponentBase } from "@actionishope/shelley";
 
 import { st, classes } from "./previewModes.st.css";
-import type { ShelleyComponentBase } from "../../../shared-types";
+import type { ShelleyComponentBase } from "../../typings/shared-types";
 
 export type PreviewModeType = "web" | "mobile" | "tablet" | "laptop";
