+++ tsconfig.json
@@ -74,6 +74,6 @@
   },
   // https://stackoverflow.com/questions/58999086/cypress-causing-type-errors-in-jest-assertions
   // https://github.com/cypress-io/cypress-and-jest-typescript-example
-  "include": ["src/**/*", "shared-types.d.ts"],
+  "include": ["src/**/*"],
   "exclude": ["cypress.d.ts"]
 }
